+++ tests/libtest/cli_hx_download.c
@@ -220,7 +220,7 @@ static int my_progress_d_cb(void *userdata,
 #endif
       default:
         curl_mfprintf(stderr, "[t-%zu] info SSL_PTR backend=%d, ptr=%p\n",
-                      t->idx, tls->backend, tls->internals);
+                      t->idx, (int)tls->backend, tls->internals);
         break;
       }
     }
